requirements.txt
numpy
pandas
scikit-learn
torch
torchvision
transformers
wandb
opencv-python
Pillow
tqdm
optuna
hydra
timm